experiments/10_ebt_vb/alexiglad_EBT/requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.0
diffusers==0.32.1
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lightning-utilities==0.11.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
nltk==3.9.1
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.2
protobuf==5.28.1
psutil==6.0.0
pyarrow==17.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.15.2
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sympy==1.13.2
timm==1.0.15
tokenizers==0.19.1
torch==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.18.0
wheel==0.44.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.11.1
zipp==3.21.0
